Add files via upload
In my previous role at CloudCard Inc, I was tasked with enhancing our fraud detection system to better identify and mitigate high-risk activities. This involved the development and implementation of a set of rules designed to detect and prevent fraudulent activities effectively. The project aimed to increase the security and reliability of our transactions and protect both the company and our customers from potential fraud.

Objective: The primary objective of this project was to create a robust set of high-risk activity rules to identify and flag suspicious transactions and behaviors, reducing the incidence of fraud within our system.

Key Components:

Rule Development: I developed a comprehensive set of rules that focused on identifying patterns and behaviors indicative of fraudulent activities. These rules were based on historical data analysis and industry best practices.

Implementation: The rules were integrated into the existing fraud detection system. This involved close collaboration with the IT and data analytics teams to ensure seamless integration and optimal performance.